% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_trips_days}
\alias{get_trips_days}
\title{Quickly pull up text pipeline for using fst database with daily data}
\usage{
get_trips_days()
}
\description{
This function allows you quickly access the pipeline necesarry to pull daily trip files through the fst database.
}
\examples{
get_trips_days()
}
\keyword{get}
