% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_trips_days_p}
\alias{get_trips_days_p}
\title{Quickly pull up text pipeline for using fst database with daily data in parallel}
\usage{
get_trips_days_p()
}
\arguments{
\item{cl}{the number of cores you want}
}
\description{
This function allows you quickly access the pipeline necesarry to pull daily trip files through the fst database in parallel.
}
\examples{
get_trips_days_p()
}
\keyword{get}
