% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{spatial_join}
\alias{spatial_join}
\title{Spatial Join function to quickly run a spatial join}
\usage{
spatial_join(data_set, shp_file, projection, long, lat)
}
\arguments{
\item{data_set}{the data set name no quotes}

\item{shp_file}{the shape file name no quotes}

\item{projection}{the projection to transform the shape file in quotes}

\item{long}{longitude}

\item{lat}{latitude}
}
\description{
This function allows you to quickly run spatial joins
}
\examples{
 spatial_join(bc, zones ,"+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs" , "longitude" , "latitude")
}
\keyword{spatial}
